% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform.R
\docType{package}
\name{platform-package}
\alias{platform-package}
\title{platform: A package to render QC reports for research groups at SII.}
\description{
platform: A package to render QC reports for research groups at SII.
}
