% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{report_bulk}
\alias{report_bulk}
\title{Render Bulk RNA-seq Report}
\usage{
report_bulk(
  dir_out,
  name_out = "QC_bulk.html",
  files_meta,
  files_meta2,
  dir_qc,
  dir_qc_fastq,
  multiqc = "multiqc",
  zip = F,
  ...
)
}
\arguments{
\item{dir_out}{the directory of the report.}

\item{name_out}{the name of the report, default: QC_bulk.html.}

\item{files_meta}{a list, first meta file.}

\item{files_meta2}{a list, second meta file.}

\item{dir_qc}{the top/father directory of QC figures.}

\item{dir_qc_fastq}{the top/father directory of FastQC results.}

\item{multiqc}{to specify an executable multiqc, default: multiqc.}

\item{zip}{if use tar to zip the out directory, default: FALSE.}

\item{...}{out name of the zipped file, default: name = "report".}
}
\description{
This function is to generate bulk Report.
}
\examples{
\dontrun{
report_bulk(
  dir_out = "./", files_meta = "/my/meta/", files_meta2 = "/my/meta2/",
  dir_qc = "/my/qc/", dir_qc_fastq = "/my/fastqc/"
)
}
}
