% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{report_sc}
\alias{report_sc}
\title{Render Single-Cell RNA-seq Report}
\usage{
report_sc(dir_out, name_out = "QC_sc.html", dir_html, dir_qc, zip = F, ...)
}
\arguments{
\item{dir_out}{directory of the report.}

\item{name_out}{name of the report, default: QC_sc.html.}

\item{dir_html}{directory of cell ranger outputs.}

\item{dir_qc}{directory of QC figures.}

\item{zip}{if use tar to zip the out directory, default: FALSE.}

\item{...}{out name of the zipped file, default: report.}
}
\description{
This function is to generate scReport.
}
\examples{
\dontrun{
report_sc(dir_out = "./", dir_html = "/my/html/", dir_qc = "/my/qc/")
}
}
