% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{report_tt}
\alias{report_tt}
\title{Render Single-Cell RNA-seq Report (Total-seq)}
\usage{
report_tt(
  dir_out,
  name_out = "QC_tt.html",
  dir_html,
  ref_csv,
  dir_qc,
  zip = F,
  ...
)
}
\arguments{
\item{dir_out}{the directory of the report.}

\item{name_out}{the name of the report, default: QC_tt.html.}

\item{dir_html}{the top/father directory of cell ranger outputs.}

\item{ref_csv}{the path to the feature_ref file.}

\item{dir_qc}{the top/father directory of QC figures.}

\item{zip}{if use tar to zip the out directory, default: FALSE.}

\item{...}{out name of the zipped file, default: name = "report".}
}
\description{
This function is to generate scReport (Total-seq).
}
\examples{
\dontrun{
report_tt(
  dir_out = "./", dir_html = "/my/html/", ref_csv = "/my/ref_feature.csv",
  dir_qc = "/my/qc/"
)
}
}
